+++ aldrin/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `UnknownCall`, which represents an unknown pending call.
 - Add `UnknownEvent`, which represents an unknown event.
 - Add variant `Error::InvalidEvent` and inner type `InvalidEvent`.
+- Add `Call` as a high-level alternative to `low_level::Call`.
 
 ### Changed