+++ aldrin/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.11.0] - 2025-01-07
 
 ### Added
 
@@ -247,6 +247,7 @@ This version still connects to brokers using protocol version 14.
 
 - Initial release.
 
+[0.11.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.11.0
 [0.10.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.10.0
 [0.9.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.9.0
 [0.8.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.8.0